The 2 main changes:
changed the simulation object so that each node has  dimension of L x deltaX x thicknes, so now each node is a flat and long horizontal slice of the profile. This was necessary because of the constant Q condition. The cosntant Q condition supplies power to all the surface. With the previous model, where we were using a node of volume thicknes x deltaX x deltaX instead of L x deltaX x thickness, then the Q would have to be divided appropriately to fit this thin vertical slice.

MOST IMPORTANTLY:
Found the constant Q bug and fixed it. The constant Q condition was fixing the output table instead of just adding to it.